+++ CHANGELOG.md
@@ -55,10 +55,12 @@ All notable changes to this project will be documented in this file.
   [recommended by Kotlin][directory-structure-convention] (issue [#260]).
 - Align styles of version [4.2.0] with the latest ones in [API reference] (issue
   [#261]).
+- Opt-in message of experimental annotations (PR [#328]).
 
 [4.2.0]: https://github.com/kotools/types/releases/tag/4.2.0
 [#260]: https://github.com/kotools/types/issues/260
 [#261]: https://github.com/kotools/types/issues/261
+[#328]: https://github.com/kotools/types/pull/328
 [directory-structure-convention]: https://kotlinlang.org/docs/coding-conventions.html#directory-structure
 
 ### Fixed